added empty spring application to project
