+++ src/main/java/com/currencyconverter/CurrencyConverterSpringApplication.java
@@ -1,0 +1,11 @@
+package com.currencyconverter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CurrencyConverterSpringApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(CurrencyConverterSpringApplication.class, args);
+	}
+}
