--- src/main/java/com/currencyconverter/com/currencyconverter/maven/rates.txt
@@ -1,2 +1,0 @@
-EUR,USD,not a number
-USD,EUR,0.90
