+++ src/test/java/com/currencyconverter/CurrencyconverterApplicationTests.java
@@ -1,0 +1,13 @@
+package com.currencyconverter;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CurrencyconverterApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
